add documentation

+++ src/core/simplify.rs
@@ -8,11 +8,11 @@ use crate::core::overlay::{Overlay, ShapeType};
 use crate::core::overlay_rule::OverlayRule;
 
 pub trait Simplify {
-    fn simplify(&self, fill_rule: FillRule, min_area: i64) -> Vec<IntShape>;
+    fn simplify(&self, fill_rule: FillRule, min_area: usize) -> Vec<IntShape>;
 }
 
 impl Simplify for IntPath {
-    fn simplify(&self, fill_rule: FillRule, min_area: i64) -> Vec<IntShape> {
+    fn simplify(&self, fill_rule: FillRule, min_area: usize) -> Vec<IntShape> {
         let mut overlay = Overlay::new(self.len());
         overlay.add_path(self, ShapeType::Subject);
 
@@ -22,7 +22,7 @@ impl Simplify for IntPath {
 }
 
 impl Simplify for [IntPath] {
-    fn simplify(&self, fill_rule: FillRule, min_area: i64) -> Vec<IntShape> {
+    fn simplify(&self, fill_rule: FillRule, min_area: usize) -> Vec<IntShape> {
         let mut overlay = Overlay::new(self.points_count());
 
         overlay.add_paths(self, ShapeType::Subject);
@@ -33,7 +33,7 @@ impl Simplify for [IntPath] {
 }
 
 impl Simplify for IntShape {
-    fn simplify(&self, fill_rule: FillRule, min_area: i64) -> Vec<IntShape> {
+    fn simplify(&self, fill_rule: FillRule, min_area: usize) -> Vec<IntShape> {
         let mut overlay = Overlay::new(self.points_count());
         overlay.add_shape(self, ShapeType::Subject);
         let graph = overlay.into_graph(fill_rule);
@@ -42,7 +42,7 @@ impl Simplify for IntShape {
 }
 
 impl Simplify for [IntShape] {
-    fn simplify(&self, fill_rule: FillRule, min_area: i64) -> Vec<IntShape> {
+    fn simplify(&self, fill_rule: FillRule, min_area: usize) -> Vec<IntShape> {
         let mut overlay = Overlay::new(self.len());
         for shape in self.iter() {
             overlay.add_shape(shape, ShapeType::Subject);
